api/pyproject.toml
[project]
name = "liiaa-api"
version = "0.1.0"
description = "Backend services of liiaa."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "openai>=1.63.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
